Completed budget expense tracking with respect to budget in the budget page
Now, expenses are summed every trip according to their respective categories, and this spending is reflected in the budget progress bars, so now we can clearly see how much of the budget is used up.
Slight changes in frontend side have been made, to ensure a more user friendly experience on the budget page. Mainly, the amount spent that is indicated inside the progress bar, will be on the right of the end of the progress bar if the amount spent is too small a percentage (<25%) of the amount allocated, so this provides better visibility.